feat: overhaul settings page with modular components for a cleaner UI
- Refactored the Settings page to utilize dedicated components for Profile, Appearance, Notifications, and Security, enhancing maintainability and readability.
- Introduced new `AppearanceCard`, `NotificationsCard`, `ProfileCard`, and `SecurityCard` components, each encapsulating their respective functionalities.
- Streamlined user experience by ensuring loading states and error handling are consistently managed across the new components.

Because a well-organized settings page is like a well-organized closet—everything in its place! 🎤✨
