+++ src/pages/Settings/SecurityCard.tsx
@@ -1,0 +1,91 @@
+import { useTranslations } from '@/hooks/useTranslations';
+import { Button } from '@/components/ui/button';
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from '@/components/ui/card';
+import { Input } from '@/components/ui/input';
+import { Label } from '@/components/ui/label';
+import { Shield } from 'lucide-react';
+import { useState } from 'react';
+import { AuthService } from '@/services/auth';
+import type { UpdatePasswordRequest } from '@/types/user';
+import { toast } from 'sonner';
+
+export function SecurityCard() {
+  const { t } = useTranslations();
+  const [currentPassword, setCurrentPassword] = useState('');
+  const [newPassword, setNewPassword] = useState('');
+  const [confirmPassword, setConfirmPassword] = useState('');
+  const [isUpdatingPassword, setIsUpdatingPassword] = useState(false);
+
+  const handlePasswordUpdate = async () => {
+    if (newPassword !== confirmPassword) {
+      toast.error('Passwords do not match');
+      return;
+    }
+    const payload: UpdatePasswordRequest = { currentPassword, newPassword };
+    setIsUpdatingPassword(true);
+    try {
+      await AuthService.updatePassword(payload);
+      setCurrentPassword('');
+      setNewPassword('');
+      setConfirmPassword('');
+      toast.success('Password updated');
+    } catch {
+      toast.error('Something went wrong');
+    } finally {
+      setIsUpdatingPassword(false);
+    }
+  };
+
+  return (
+    <Card>
+      <CardHeader>
+        <CardTitle className="flex items-center">
+          <Shield className="mr-2 h-5 w-5" />
+          {t('settings.security.title')}
+        </CardTitle>
+        <CardDescription>{t('settings.security.description')}</CardDescription>
+      </CardHeader>
+      <CardContent className="space-y-4">
+        <div className="space-y-2">
+          <Label htmlFor="current-password">Current Password</Label>
+          <Input
+            id="current-password"
+            type="password"
+            value={currentPassword}
+            onChange={e => setCurrentPassword(e.target.value)}
+            disabled={isUpdatingPassword}
+          />
+        </div>
+        <div className="space-y-2">
+          <Label htmlFor="new-password">New Password</Label>
+          <Input
+            id="new-password"
+            type="password"
+            value={newPassword}
+            onChange={e => setNewPassword(e.target.value)}
+            disabled={isUpdatingPassword}
+          />
+        </div>
+        <div className="space-y-2">
+          <Label htmlFor="confirm-password">Confirm Password</Label>
+          <Input
+            id="confirm-password"
+            type="password"
+            value={confirmPassword}
+            onChange={e => setConfirmPassword(e.target.value)}
+            disabled={isUpdatingPassword}
+          />
+        </div>
+        <Button onClick={handlePasswordUpdate} disabled={isUpdatingPassword}>
+          {isUpdatingPassword ? t('common.saving') : t('auth.updatePassword')}
+        </Button>
+      </CardContent>
+    </Card>
+  );
+}
